Drop support for job options passed to `Que::Job.enqueue` outside the `job_options` keyword argument
This was deprecated in 1.2.0, as it's difficult to support once we split the args and kwargs for Ruby 3.

This change passes all the regular specs, but unfortunately not the Rails ones, due to being unable to distinguish a hash from kwargs. That'll have to wait to be fixed by kwargs splitting.